homewotk7: fix after review
